Split validation tests and and invalid test
